+++ apps/website/content/docs/changelog.md
@@ -2,14 +2,24 @@
 title: Changelog
 ---
 
+## v2.9.0 (2026-02-02)
+
+### ✨ New
+
+- Rename `no-non-async-server-functions` to `rsc-no-misused-use-server`
+
+### 📝 Changes you should be aware of
+
+The `no-non-async-server-functions` rule has been renamed to `rsc-no-misused-use-server`. Please update your configuration accordingly if you are using it.
+
 ## v2.8.4 (2026-02-01)
 
 ### 🐞 Fixes
 
 - Change status of `jsx-shorthand-boolean` rule in migration table by @connorshea in https://github.com/Rel1cx/eslint-react/pull/1446
 - Fix rule name in docs for `debug/is-from-ref` by @amir-rahmanii in https://github.com/Rel1cx/eslint-react/pull/1445
-- Fix rule name in docs for `no-non-async-server-functions` by @Rel1cx in https://github.com/Rel1cx/eslint-react/pull/1447
-- Mark `no-non-async-server-functions` as experimental by @Rel1cx in https://github.com/Rel1cx/eslint-react/pull/1448
+- Fix rule name in docs for `rsc-no-misused-use-server` by @Rel1cx in https://github.com/Rel1cx/eslint-react/pull/1447
+- Mark `rsc-no-misused-use-server` as experimental by @Rel1cx in https://github.com/Rel1cx/eslint-react/pull/1448
 
 ## New Contributors
 
@@ -21,7 +31,7 @@ title: Changelog
 
 ### ✨ New
 
-- feat(#1435): implements `no-non-async-server-functions` by @SukkaW in https://github.com/Rel1cx/eslint-react/pull/1443
+- feat(#1435): implements `rsc-no-misused-use-server` by @SukkaW in https://github.com/Rel1cx/eslint-react/pull/1443
 
 **Full Changelog**: https://github.com/Rel1cx/eslint-react/compare/v2.8.1...v2.8.3
