+++ apps/website/content/docs/rules/meta.json
@@ -39,7 +39,6 @@
     "no-misused-capture-owner-stack",
     "no-nested-component-definitions",
     "no-nested-lazy-component-declarations",
-    "no-non-async-server-functions",
     "no-prop-types",
     "no-redundant-should-component-update",
     "no-set-state-in-component-did-mount",
@@ -66,6 +65,7 @@
     "prefer-namespace-import",
     "prefer-read-only-props",
     "prefer-use-state-lazy-initialization",
+    "rsc-no-misused-use-server",
     "---DOM Rules---",
     "dom-no-dangerously-set-innerhtml",
     "dom-no-dangerously-set-innerhtml-with-children",
