+++ apps/website/content/docs/rules/overview.mdx
@@ -64,7 +64,6 @@ full: true
 | [`no-misused-capture-owner-stack`](no-misused-capture-owner-stack)                 | 0️⃣ 2️⃣ |    `🧪`   | Prevents incorrect usage of `captureOwnerStack`                                                                               |          |
 | [`no-nested-component-definitions`](no-nested-component-definitions)               | 2️⃣ 2️⃣ |           | Disallows nesting component definitions inside other components                                                               |          |
 | [`no-nested-lazy-component-declarations`](no-nested-lazy-component-declarations)   | 2️⃣ 2️⃣ |           | Disallows nesting lazy component declarations inside other components                                                         |          |
-| [`no-non-async-server-functions`](no-non-async-server-functions)                   | 0️⃣ 0️⃣ | `🔧` `🧪` | Disallows non-async [React Server Functions](https://react.dev/reference/rsc/server-functions)                                |          |
 | [`no-prop-types`](no-prop-types)                                                   | 2️⃣ 2️⃣ |           | Disallows `propTypes` in favor of TypeScript or another type-checking solution                                                |          |
 | [`no-redundant-should-component-update`](no-redundant-should-component-update)     | 2️⃣ 2️⃣ |           | Disallows `shouldComponentUpdate` when extending `React.PureComponent`                                                        |          |
 | [`no-set-state-in-component-did-mount`](no-set-state-in-component-did-mount)       | 1️⃣ 1️⃣ |           | Disallows calling `this.setState` in `componentDidMount` outside functions such as callbacks                                  |          |
@@ -74,8 +73,8 @@ full: true
 | [`no-unnecessary-key`](no-unnecessary-key)                                         | 0️⃣ 0️⃣ |    `🧪`   | Disallows unnecessary `key` props on nested child elements when rendering lists                                               |          |
 | [`no-unnecessary-use-callback`](no-unnecessary-use-callback)                       | 0️⃣ 1️⃣ |    `🧪`   | Disallows unnecessary usage of `useCallback`                                                                                  |          |
 | [`no-unnecessary-use-memo`](no-unnecessary-use-memo)                               | 0️⃣ 1️⃣ |    `🧪`   | Disallows unnecessary usage of `useMemo`                                                                                      |          |
-| [`no-unnecessary-use-ref`](no-unnecessary-use-ref)                                 | 0️⃣ 0️⃣ |    `🧪`   | Disallows unnecessary usage of `useRef`                                                                                       |          |
 | [`no-unnecessary-use-prefix`](no-unnecessary-use-prefix)                           | 1️⃣ 1️⃣ |           | Enforces that a function with the `use` prefix uses at least one Hook inside it                                               |          |
+| [`no-unnecessary-use-ref`](no-unnecessary-use-ref)                                 | 0️⃣ 0️⃣ |    `🧪`   | Disallows unnecessary usage of `useRef`                                                                                       |          |
 | [`no-unsafe-component-will-mount`](no-unsafe-component-will-mount)                 | 1️⃣ 1️⃣ |           | Warns about the use of `UNSAFE_componentWillMount` in class components                                                        |          |
 | [`no-unsafe-component-will-receive-props`](no-unsafe-component-will-receive-props) | 1️⃣ 1️⃣ |           | Warns about the use of `UNSAFE_componentWillReceiveProps` in class components                                                 |          |
 | [`no-unsafe-component-will-update`](no-unsafe-component-will-update)               | 1️⃣ 1️⃣ |           | Warns about the use of `UNSAFE_componentWillUpdate` in class components                                                       |          |
@@ -91,6 +90,7 @@ full: true
 | [`prefer-namespace-import`](prefer-namespace-import)                               | 0️⃣ 0️⃣ |    `🔧`   | Enforces importing React via a namespace import                                                                               |          |
 | [`prefer-read-only-props`](prefer-read-only-props)                                 | 0️⃣ 0️⃣ | `💭` `🧪` | Enforces read-only props in components                                                                                        |          |
 | [`prefer-use-state-lazy-initialization`](prefer-use-state-lazy-initialization)     | 1️⃣ 1️⃣ |           | Enforces wrapping function calls made inside `useState` in an `initializer function`                                          |          |
+| [`rsc-no-misused-use-server`](rsc-no-misused-use-server)                           | 0️⃣ 0️⃣ | `🔧` `🧪` | Disallows non-async [React Server Functions](https://react.dev/reference/rsc/server-functions)                                |          |
 
 ## DOM Rules
